# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PRAYAG eCommerce Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/apply-coupon:
    post:
      operationId: applyCoupon
      tags: [cart]
      summary: Apply coupon to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "200":
          description: Cart with discount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # Wishlist
...
  /admin/products:
    get:
      operationId: listAdminProducts
      tags: [admin]
      summary: Admin product list
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [admin]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
  /admin/products/{id}:
    patch:
      operationId: updateProduct
      tags: [admin]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
    delete:
      operationId: deleteProduct
      tags: [admin]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: All orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /admin/orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /admin/customers:
    get:
      operationId: listAdminCustomers
      tags: [admin]
      summary: All customers (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
  /admin/dealers:
    get:
      operationId: listAdminDealers
      tags: [admin]
      summary: All dealers (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Dealers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dealer"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    UserRegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        role:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
...
    Product:
      type: object
      required: [id, name, slug, sku, price, mrp, categoryId, rating, reviewCount, inStock]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        sku:
          type: string
        price:
          type: number
        mrp:
          type: number
        discount:
          type: ["number", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        inStock:
          type: boolean
        isFeatured:
          type: boolean
        isNew:
          type: boolean
    ProductDetail:
      type: object
      required: [id, name, slug, sku, price, mrp, categoryId, rating, reviewCount, inStock, images, description]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        sku:
          type: string
        price:
          type: number
        mrp:
          type: number
        discount:
          type: ["number", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        imageUrl:
          type: ["string", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        inStock:
          type: boolean
        description:
          type: string
        specifications:
          type: ["string", "null"]
        warranty:
          type: ["string", "null"]
        gstPercent:
          type: number
        isFeatured:
          type: boolean
        isNew:
          type: boolean
    ProductListResponse:
      type: object
      required: [products, total, page, totalPages]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
    ProductInput:
      type: object
      required: [name, sku, price, mrp, categoryId, description]
      properties:
        name:
          type: string
        sku:
          type: string
        price:
          type: number
        mrp:
          type: number
        categoryId:
          type: integer
        description:
          type: string
        specifications:
          type: string
        warranty:
          type: string
        gstPercent:
          type: number
        isFeatured:
          type: boolean
        isNew:
          type: boolean
        imageUrl:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        mrp:
          type: number
        description:
          type: string
        isFeatured:
          type: boolean
        isNew:
          type: boolean
        inStock:
          type: boolean
    CartItem:
      type: object
      required: [id, productId, productName, productSlug, imageUrl, price, quantity, subtotal]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productSlug:
          type: string
        imageUrl:
          type: ["string", "null"]
        price:
          type: number
        quantity:
          type: integer
        subtotal:
          type: number
    Cart:
      type: object
      required: [items, subtotal, gst, shipping, total, itemCount]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        gst:
          type: number
        shipping:
          type: number
        discount:
          type: number
        couponCode:
          type: ["string", "null"]
        total:
          type: number
        itemCount:
          type: integer
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
    CouponInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
...
    Address:
      type: object
      required: [id, name, phone, street, city, state, pincode, country]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        country:
          type: string
        isDefault:
          type: boolean
...
    OrderItem:
      type: object
      required: [id, productId, productName, imageUrl, price, quantity, subtotal]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        imageUrl:
          type: ["string", "null"]
        price:
          type: number
        quantity:
          type: integer
        subtotal:
          type: number
    Order:
      type: object
      required: [id, orderNumber, status, items, subtotal, gst, shipping, total, createdAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        gst:
          type: number
        shipping:
          type: number
        discount:
          type: number
        total:
          type: number
        paymentMethod:
          type: ["string", "null"]
        paymentStatus:
          type: ["string", "null"]
        shippingAddress:
          $ref: "#/components/schemas/Address"
        createdAt:
          type: string
        updatedAt:
          type: string
...
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
...
    Customer:
      type: object
      required: [id, name, email, createdAt, totalOrders]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        totalOrders:
          type: integer
        totalSpent:
          type: number
        createdAt:
          type: string
    Dealer:
      type: object
      required: [id, businessName, contactName, email, status, createdAt]
      properties:
        id:
          type: integer
        businessName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        status:
          type: string
        totalOrders:
          type: integer
        createdAt:
          type: string
...