# lib/api-zod/src/generated/types/dealer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PRAYAG eCommerce Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Dealer {
  id: number;
  businessName: string;
  contactName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  status: string;
  totalOrders?: number;
  createdAt: string;
}