# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PRAYAG eCommerce Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Address } from './address';
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  orderNumber: string;
  status: string;
  items: OrderItem[];
  subtotal: number;
  gst: number;
  shipping: number;
  discount?: number;
  total: number;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  paymentStatus?: string | null;
  shippingAddress?: Address;
  createdAt: string;
  updatedAt?: string;
}