# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PRAYAG eCommerce Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  slug: string;
  sku: string;
  price: number;
  mrp: number;
  /** @nullable */
  discount?: number | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  rating: number;
  reviewCount: number;
  inStock: boolean;
  isFeatured?: boolean;
  isNew?: boolean;
}