# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PRAYAG eCommerce Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  sku: string;
  price: number;
  mrp: number;
  categoryId: number;
  description: string;
  specifications?: string;
  warranty?: string;
  gstPercent?: number;
  isFeatured?: boolean;
  isNew?: boolean;
  imageUrl?: string;
}