# lib/api-zod/src/generated/types/userRegisterInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PRAYAG eCommerce Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserRegisterInput {
  name: string;
  email: string;
  password: string;
  phone?: string;
  role?: string;
}